updated util DateTime functions to be fieldName-agnostic, split isMinor into separate method, moved util files to a more consistent location
