+++ app/javascript/components/patient/household/HouseholdMemberTable.js
@@ -2,11 +2,11 @@ import React from 'react';
 import { PropTypes } from 'prop-types';
 import moment from 'moment-timezone';
 import _ from 'lodash';
-import { formatNameAlt, formatDateOfBirth } from '../../../../utils/Patient';
+import { formatNameAlt, formatDateOfBirth } from '../../../utils/Patient';
 
-import BadgeHoH from '../utils/BadgeHoH';
-import CustomTable from '../../../layout/CustomTable';
 import { patientHref } from '../../../../utils/Navigation';
+import BadgeHoH from '../icons/BadgeHoH';
+import CustomTable from '../../layout/CustomTable';
 
 class HouseholdMemberTable extends React.Component {
   constructor(props) {
@@ -15,7 +15,7 @@ class HouseholdMemberTable extends React.Component {
       table: {
         colData: [
           { field: 'name', label: 'Name', isSortable: true, tooltip: null, filter: this.formatPatientName },
-          { field: 'date_of_birth', label: 'Date of Birth', isSortable: true, tooltip: null, filter: formatDateOfBirth },
+          { field: 'date_of_birth', label: 'Date of Birth', isSortable: true, tooltip: null, filter: this.formatDOB },
           { field: 'isolation', label: 'Workflow', isSortable: true, tooltip: null, options: { true: 'Isolation', false: 'Exposure' } },
           {
             field: 'monitoring',
@@ -269,6 +269,11 @@ class HouseholdMemberTable extends React.Component {
         </a>
       </React.Fragment>
     );
+  };
+
+  formatDOB = data => {
+    const rowData = data.rowData;
+    return formatDateOfBirth(rowData.date_of_birth, rowData.id);
   };
 
   /**
