+++ app/javascript/components/patient/icons/BadgeHoH.js
