+++ app/javascript/components/patient/icons/IconMinor.js
