+++ app/javascript/tests/patient/Patient.test.js
@@ -5,8 +5,8 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import _ from 'lodash';
 import Patient from '../../components/patient/Patient';
 import FollowUpFlagPanel from '../../components/patient/follow_up_flag/FollowUpFlagPanel';
-import BadgeHoH from '../../components/patient/household/utils/BadgeHoH';
 import InfoTooltip from '../../components/util/InfoTooltip';
+import BadgeHoH from '../../components/patient/icons/BadgeHoH';
 import { mockUser1 } from '../mocks/mockUsers';
 import { mockPatient1, mockPatient2, mockPatient3, mockPatient4, mockPatient5, blankMockPatient } from '../mocks/mockPatients';
 import { mockJurisdictionPaths } from '../mocks/mockJurisdiction';
