+++ app/javascript/tests/patient/icons/BadgeHoH.test.js
@@ -2,8 +2,7 @@ import React from 'react';
 import { shallow } from 'enzyme';
 import { Badge } from 'react-bootstrap';
 import ReactTooltip from 'react-tooltip';
-import BadgeHoH from '../../../../components/patient/household/utils/BadgeHoH';
-
+import BadgeHoH from '../../../components/patient/icons/BadgeHoH';
 describe('BadgeHoH', () => {
   it('Properly renders all main components', () => {
     const wrapper = shallow(<BadgeHoH patientId={'1'} location={'right'} />);
