+++ app/javascript/utils/Patient.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import IconMinor from '../components/patient/utils/IconMinor';
+import IconMinor from '../components/patient/icons/IconMinor';
 import { formatDate } from '../utils/DateTime';
 import moment from 'moment-timezone';
 import { faDatabase } from '@fortawesome/free-solid-svg-icons';
@@ -67,36 +67,30 @@ function formatRace(patient) {
   return raceArray.length === 0 ? '--' : raceArray.join(', ');
 }
 
+/**
+ * helper function to determine if a given date of birth would make someone a minor.
+ * @param {*} date : a date of birth value in YYYY-MM-DD format
+ * @returns boolean true if patient is under 18, false if not
+ */
 function isMinor(date) {
   return moment(date, 'YYYY-MM-DD').isAfter(moment().subtract(18, 'years')); 
 }
 
 /**
  * Formats values in the date of birth column to be human readable and include whether that DOB indicates a minor.
- * @param {Object} data - Data about the cell this filter is called on.
+ * @param {String} dateOfBirth - Patient's date of birth in YYYY-MM-DD format
+ * @param {Int} id - Patient'd unique ID
  */
-function formatDateOfBirth(data) {
-  const rowData = data.rowData;
-  if (!!rowData?.dob && isMinor(rowData.dob)) {
+function formatDateOfBirth(dateOfBirth, id) {
+  if (isMinor(dateOfBirth)) {
     return (
       <div>
-        <IconMinor patientId={rowData.id.toString()} customClass={'float-right ml-1'} />
-        {formatDate(rowData.dob)}
+        <IconMinor patientId={id.toString()} customClass={'float-right ml-1'} />
+        {formatDate(dateOfBirth)}
       </div>
     );
   }
-  if (!!rowData?.date_of_birth && isMinor(rowData.date_of_birth)) {
-    return (
-      <div>
-        <IconMinor patientId={rowData.id.toString()} customClass={'float-right ml-1'} />
-        {formatDate(rowData.date_of_birth)}
-      </div>
-    );
-  }
-  if (!!rowData?.date_of_birth) {
-    return formatDate(rowData.date_of_birth);
-  }
-  return formatDate(rowData.dob);
+  return formatDate(dateOfBirth);
 }
 
 export { formatName, formatNameAlt, formatPhoneNumber, formatRace, isMinor, formatDateOfBirth };
